requirements.txt
agate==1.7.1
annotated-types==0.7.0
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dbt-core==1.7.0
dbt-extractor==0.5.1
dbt-postgres==1.7.0
dbt-semantic-interfaces==0.4.4
delta-spark==3.2.1
deltalake==0.21.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
Logbook==1.5.3
MarkupSafe==3.0.2
mashumaro==3.14
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
networkx==3.4.2
packaging==24.1
parsedatetime==2.6
pathspec==0.11.2
protobuf==4.25.3
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
setuptools==75.3.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==1.26.20
zipp==3.20.2
